requirements.txt
# ── 数据采集 ──────────────────────────────────────────
httpx==0.28.1
beautifulsoup4==4.14.3
requests==2.32.3
pydantic==2.7.4
python-dotenv==1.0.1

# ── 数据库 ────────────────────────────────────────────
chromadb==0.5.5

# ── AI / ML ───────────────────────────────────────────
sentence-transformers==3.0.1
setfit==1.1.3
datasets==4.5.0
transformers==4.45.2
accelerate==1.10.1
torch==2.2.2
scikit-learn==1.6.1
numpy==1.26.4
